+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ### Changed
 
 - Don't add temp prefix when using `extract` command.
+- VS Code extension mirrord enable/disable to be per workspace.
 
 ## 3.13.3